lab1/requirements.txt
boto3==1.37.23
botocore==1.37.23
jmespath==1.0.1
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.11.4
six==1.17.0
urllib3==2.3.0